logic improvements

+++ pkg/config/remote/service/util.go
@@ -32,13 +32,19 @@ type AgentMetadata struct {
 
 func recreate(path string) (*bbolt.DB, error) {
 	log.Infof("Clear remote configuration database")
-	err := os.Remove(path)
-	if err != nil {
-		log.Debugf("failed to remove rc db: %v", err)
+	_, err := os.Stat(path)
+	if err != nil && !os.IsNotExist(err) {
+		return nil, fmt.Errorf("could not check if rc db exists: (%s): %v", path, err)
+	}
+	if err == nil {
+		err := os.Remove(path)
+		if err != nil {
+			return nil, fmt.Errorf("could not remote existing rc db (%s): %v", path, err)
+		}
 	}
 	err = os.MkdirAll(filepath.Dir(path), 0700)
 	if err != nil {
-		return nil, fmt.Errorf("failed to create rc db dir: %v", err)
+		return nil, fmt.Errorf("failed to create rc db dir: (%s): %v", path, err)
 	}
 	db, err := bbolt.Open(path, 0600, &bbolt.Options{})
 	if err != nil {
